refactor: index 파일 생성 및 import 경로 정리
