+++ frontend/src/components/VoteInterface/index.ts
@@ -1,0 +1,4 @@
+export { default as VoteInterface } from './VoteInterface';
+export { default as useVoteInterfaceContext } from './hooks/useVoteInterfaceContext';
+export { VoteInterfaceProvider } from './VoteInterfaceProvider';
+export { VoteInterfaceContext } from './VoteInterfaceProvider';
